Various updates, switch to CMake build and SDK v4.1 (latest stable release)
SDK builds in Win32 (needs to add -G Ninja explicitly)
Use sub-make to copy files after building SDK - portable
Add ESP_VARIANT support
Fix malloc_count - needs to wrap strdup as well
Fix compile warnings
Fix flash storage

Build works for pre-release IDF v4.2 but fails during compilation

+++ tests/HostTests/app/test-stream.cpp
@@ -65,7 +65,7 @@ public:
 			String s(FS_abstract);
 			MemoryDataStream stream(std::move(s));
 			TEST_ASSERT(!s);
-			TEST_ASSERT(FS_abstract.length() == stream.available());
+			TEST_ASSERT(FS_abstract.length() == size_t(stream.available()));
 			// And back again
 			stream.moveString(s);
 			TEST_ASSERT(FS_abstract == s);
@@ -75,7 +75,7 @@ public:
 		TEST_CASE("LimitedMemoryStream::moveString (1)")
 		{
 			FSTR::Stream src(FS_abstract);
-			REQUIRE(src.available() == FS_abstract.length());
+			REQUIRE(size_t(src.available()) == FS_abstract.length());
 			LimitedMemoryStream dest(FS_abstract.length());
 			// Operation will drop last character as this is converted to NUL
 			dest.copyFrom(&src);
@@ -91,7 +91,7 @@ public:
 		TEST_CASE("LimitedMemoryStream::moveString (2)")
 		{
 			FSTR::Stream src(FS_abstract);
-			REQUIRE(src.available() == FS_abstract.length());
+			REQUIRE(size_t(src.available()) == FS_abstract.length());
 			LimitedMemoryStream dest(FS_abstract.length() + 10);
 			// Operation will drop last character as this is converted to NUL
 			dest.copyFrom(&src);
